Overhauled the structure of src/ to make the package more modular with packages and subpackages.
